sending&saving a message

+++ app/src/main/java/com/beyond/amber/ChatData.java
@@ -5,5 +5,5 @@ import java.io.Serializable;
 public class ChatData implements Serializable {
     public String msg;
     public int sender;
-    public int time;
+    public long time;
 }
